requirements.txt
fastapi[all]
python-jose[cryptography]
passlib[bcrypt]
numpy
openai==0.27.4
retry==0.9.2





